Refactored audio download
